Throw out half-baked support for global variables
This change reverts commit 83370a5, which was too short-sighted. I
wanted to make SSA minimization work in the presence of variables that
are live on entry to the first basic block. The problem is that
variables, if live on entry to any basic block, may be assigned SSA
names that are indistinguishable from SSA names of variables that are
live on entry to the first basic block. In other words, when looking at
a phi-function like `x_0 := PHI(x_-1, x_2)`, we have not enough context
to decide whether `x_-1` must be kept around or thrown away to enable
replacing `x_0` with `x_2`.

In the interest of keeping things simple, global variables are out, and
every variable that is used before being defined must become a parameter
of the first basic block.