Label entry basic blocks
From the LLVM Language Reference Manual: "Unnamed temporaries are
numbered sequentially (using a per-function incrementing counter,
starting with 0). Note that basic blocks and unnamed function parameters
are included in this numbering. For example, if the entry basic block is
not given a label name and all function parameters are named, then it
will get number 0."